+++ keras_adamw/optimizers_225.py
@@ -160,9 +160,9 @@ class AdamW(Optimizer):
             'lr_multipliers': self.lr_multipliers,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K.get_value(self.t_cur)),
-            'eta_t': int(K.eval(self.eta_t)),
-            'eta_min': int(K.get_value(self.eta_min)),
-            'eta_max': int(K.get_value(self.eta_max)),
+            'eta_t': float(K.eval(self.eta_t)),
+            'eta_min': float(K.get_value(self.eta_min)),
+            'eta_max': float(K.get_value(self.eta_max)),
             'init_verbose': self.init_verbose,
             'epsilon': self.epsilon,
             'amsgrad': self.amsgrad
@@ -304,9 +304,9 @@ class NadamW(Optimizer):
             'lr_multipliers': self.lr_multipliers,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K.get_value(self.t_cur)),
-            'eta_t': int(K.eval(self.eta_t)),
-            'eta_min': int(K.get_value(self.eta_min)),
-            'eta_max': int(K.get_value(self.eta_max)),
+            'eta_t': float(K.eval(self.eta_t)),
+            'eta_min': float(K.get_value(self.eta_min)),
+            'eta_max': float(K.get_value(self.eta_max)),
             'init_verbose': self.init_verbose
         }
         base_config = super(NadamW, self).get_config()
@@ -422,9 +422,9 @@ class SGDW(Optimizer):
             'lr_multipliers': self.lr_multipliers,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K.get_value(self.t_cur)),
-            'eta_t': int(K.eval(self.eta_t)),
-            'eta_min': int(K.get_value(self.eta_min)),
-            'eta_max': int(K.get_value(self.eta_max)),
+            'eta_t': float(K.eval(self.eta_t)),
+            'eta_min': float(K.get_value(self.eta_min)),
+            'eta_max': float(K.get_value(self.eta_max)),
             'init_verbose': self.init_verbose
         }
         base_config = super(SGDW, self).get_config()
