+++ keras_adamw/optimizers_v2.py
@@ -260,9 +260,9 @@ class AdamW(OptimizerV2):
             'weight_decays': self.weight_decays,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K_eval(self.t_cur)),
-            'eta_t': int(K_eval(self.eta_t)),
-            'eta_min': int(K_eval(self.eta_min)),
-            'eta_max': int(K_eval(self.eta_max)),
+            'eta_t': float(K_eval(self.eta_t)),
+            'eta_min': float(K_eval(self.eta_min)),
+            'eta_max': float(K_eval(self.eta_max)),
             'init_verbose': self.init_verbose
         })
         return config
@@ -511,9 +511,9 @@ class NadamW(OptimizerV2):
             'weight_decays': self.weight_decays,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K_eval(self.t_cur)),
-            'eta_t': int(K_eval(self.eta_t)),
-            'eta_min': int(K_eval(self.eta_min)),
-            'eta_max': int(K_eval(self.eta_max)),
+            'eta_t': float(K_eval(self.eta_t)),
+            'eta_min': float(K_eval(self.eta_min)),
+            'eta_max': float(K_eval(self.eta_max)),
             'init_verbose': self.init_verbose
         })
         return config
@@ -678,9 +678,9 @@ class SGDW(OptimizerV2):
             'weight_decays': self.weight_decays,
             'use_cosine_annealing': self.use_cosine_annealing,
             't_cur': int(K_eval(self.t_cur)),
-            'eta_t': int(K_eval(self.eta_t)),
-            'eta_min': int(K_eval(self.eta_min)),
-            'eta_max': int(K_eval(self.eta_max)),
+            'eta_t': float(K_eval(self.eta_t)),
+            'eta_min': float(K_eval(self.eta_min)),
+            'eta_max': float(K_eval(self.eta_max)),
             'init_verbose': self.init_verbose
         })
         return config
